requirements.txt
django
wsgiref
gunicorn
psycopg2
dj-database-url
newrelic
django-taggit
